internal document service
