+++ projects/valtimo/config/src/lib/services/base-api.service.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
+ *
+ * Licensed under EUPL, Version 1.2 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {HttpClient} from '@angular/common/http';
+import {ConfigService} from './config.service';
+import {UrlUtils} from '../utils';
+
+export abstract class BaseApiService {
+  private readonly _valtimoEndpointUri!: string;
+
+  constructor(
+    protected readonly httpClient: HttpClient,
+    protected readonly configService: ConfigService
+  ) {
+    this._valtimoEndpointUri = configService.config.valtimoApi.endpointUri;
+  }
+
+  public getApiUrl(urlPart: string): string {
+    return `${UrlUtils.formatUrlTrailingSlash(this._valtimoEndpointUri, false)}${urlPart}`;
+  }
+}
