+++ projects/valtimo/config/src/public-api.ts
@@ -22,10 +22,7 @@ export * from './lib/config.module';
 export * from './lib/constants';
 export * from './lib/extension/extension.component';
 export * from './lib/models';
-export * from './lib/services/config.service';
-export * from './lib/services/menu-include.service';
-export * from './lib/services/user-settings.service';
-export * from './lib/services/localization.service';
+export * from './lib/services';
 export * from './lib/translation';
 export * from './lib/utils';
 export * from './lib/testing';
