+++ projects/valtimo/document/src/lib/models/index.ts
@@ -17,3 +17,6 @@
 export * from './document.model';
 export * from './list-sorting.model';
 export * from './audit.model';
+export * from './advanced-document-search-request';
+export * from './document-search-request';
+export * from './internal-document-status.model';
