+++ projects/valtimo/document/src/lib/services/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,5 @@
  * limitations under the License.
  */
 
-/*
- * Public API Surface of document
- */
-
-export * from './lib/models';
-export * from './lib/services';
-export * from './lib/document.module';
+export * from './document-status.service';
+export * from './document.service';
