+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-detail-container/dossier-management-detail-container.component.html
@@ -89,6 +89,17 @@
         [documentDefinitionName]="documentDefinitionName$ | async"
       ></valtimo-dossier-management-tabs>
     </cds-tab>
+    <cds-tab
+      *ngIf="{active: currentTab === TabEnum.STATUSES} as vars"
+      class="no-padding-left-right no-padding-top-bottom"
+      [active]="vars.active"
+      [heading]="'dossierManagement.tabs.statuses' | translate"
+      (selected)="displayBodyComponent(TabEnum.STATUSES)"
+    >
+      <valtimo-dossier-management-statuses
+        *ngIf="vars.active"
+      ></valtimo-dossier-management-statuses>
+    </cds-tab>
   </cds-tabs>
 </ng-template>
