+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-statuses/dossier-management-statuses.component.scss
@@ -1,5 +1,5 @@
-/*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
+/*!
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/*
- * Public API Surface of document
- */
-
-export * from './lib/models';
-export * from './lib/services';
-export * from './lib/document.module';
