+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-statuses/dossier-management-statuses.component.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-/*
- * Public API Surface of document
- */
+import {Component} from '@angular/core';
 
-export * from './lib/models';
-export * from './lib/services';
-export * from './lib/document.module';
+@Component({
+  selector: 'valtimo-dossier-management-statuses',
+  templateUrl: './dossier-management-statuses.component.html',
+  styleUrls: ['./dossier-management-statuses.component.scss'],
+})
+export class DossierManagementStatusesComponent {}
